requirements.txt
django == 3.0.6
psycopg2-binary == 2.8.3
python-dotenv == 0.10.3
django-emoji ==  2.2.2
# django-emoji-picker == 0.0.6
django-admin-sso == 3.0.0
django-markdownx == 3.0.1
django-taggit == 1.3.0